1. Fixed a old aggregator function. 2. Created Archive for old code in case we still need it. 3. Archived old System 1 Conversion Aggregator 4. Archived old System 1 Sheet Values 5. Archived old System 1 Spreadsheet Controller 6. Created new System 1 Conversion Aggregator (It used to rely on other tables now it's independent) 7. Update System 1 Sheet Values 8. Created New System1 Spreadsheet Controller 9. Created New System1 Spreadsheet Column Mapper 10. Added Range Capabilites to Spreadsheet Updater Service 11. Added Documentation to System 1 Database Update Service 12. Changed the rule for Sys1 Sheet Update to every 5 min after 2nd min. 13. Created a development route. (Stashed, due to missing auth)
The old System 1 Data Updating Procedure had a bug in the SQL Query Aggregator and relied for creating conversion reports on the Campaigns, Facebook & Pb table. The problem with that was that since campagins are connected with ad accounts and not all ad accounts are stored in the datable, the campaigns that the System 1 relied to weren't in the database which produces empty conversion reports.

The new System 1 Procedure doesn't rely on any other data to update the spreadsheets. (The database was being updated just fine even with the previous version. I just added comments/documentation to how it works.) Changed the sheet to update at minutes 2, 7, 12, 17, 22, 27, 32, 37, 42, 47, 52, and 57 of every hour.