Traffic Source User Accounts (#42)
Created a User Account Service and Controller that handles all Integrated Facebook Account related data such as integrating, reauthorizing, revoking the integration, fetching user accounts from the database and so on. Every facebook account related operation throughout the application is migrated to this section of Facebook micro-service.

NOTE: The delete method is not fully tested in a live-like application. I added cascade to user accounts referenced tables when a user account is deleted but didn't had the chance to test it. This should be done in a future setup.