Feature/sedo ms (#78)
* Sedo Microservice
1. Created Sedo Service that fetched from the API.
2. Created Sedo Insights Entity that represents Sedo data.
3. Created Sedo Controller.
4. Created Sedo Repository.

Note: What's not implemented is the upsert functionality in the repository, that's also used by the Service to upsert Sedo data. The reason being that we haven't structured a Sedo Table yet, due to the fact that we are in discussions with the team over the sub params we will receive from Sedo.

* Sedo Infrastructure WIP
1. Created the sedo table migration script.
2. Created toDatabase mapper and upsert logic.

Note: The integration is stuck due to discussions on the content of the sub query params. Once we decide what they will be we will move forward to testing the integrations and integrating the ms into our unified data model.

* Added Sedo to Aggregates Compilation

* Sedo Aggregates & CRON
1. Added Sedo as an available network.
2. Added Network labeling to insights.
3. Added restrictions to hourly reports for Sedo.
4. Created Sedo Cron
5. Added Sedo Fault Handling
6. Integrated sedo to crons.