Centralized Environment Vars Manager & AWS Secrets Manager & SSM Param Store Configuration (#123)
* Environment Variable Manager
Created a centralized way of accessing environment variables. There's a switch that when set to true, fetches them from a local while, while when set to false fetches them from Secrets Manager and Parameter Store of AWS where we are supposed to fetch while in production. The access to the variables throughout the code is not yet replaced but will be so in a future commit.

* Created a server initializing script

* Repalced old accessing of env variables with new.